Look for invalid characters not part of a regex
The search must be either a string part of an ethereum address
or a regular expression.

If we have some characters that can be part of a regular expression
like backslash(\), dot(.), {}, comma(,), etc...
Then we compile the regex and use it.
Otherwise we throw an error because the string cannot match an ethereum address.

It is still possible to search for regexes that cannot match
an Ethereum address like:

[fe]+o , [fe]o, feo[], or feo\n

But _at least_ we reduce the possibility of giving a simple string
like 'feo' that cannot be part of an Ethereum address.

Closes vanieth#3